add Dockerfile
