+++ .gitignore
@@ -1,6 +1,7 @@
 ### AL ###
 #Template for AL projects for Dynamics 365 Business Central
 #launch.json folder
+.devvenv/
 .vscode/
 #Cache folder
 .alcache/
