+++ Dockerfile
@@ -1,0 +1,13 @@
+FROM python:3.11-slim
+
+WORKDIR /app
+
+COPY requirements.txt .
+
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY /src .
+
+EXPOSE 5000
+
+CMD ["flask", "--app", "app.py", "run", "--host=0.0.0.0"]
