+++ requirements.txt
@@ -1,1 +1,1 @@
-flask
+Flask==3.0.3
